application/requirements.txt
absl-py==0.12.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.6
awsebcli==3.20.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
botocore==1.20.112
bs4==0.0.1
cachetools==4.2.1
cement==2.8.2
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.1
constantly==15.1.0
crayons==0.4.0
cryptography==3.2.1
cssselect==1.1.0
datasets==1.6.0
decorator==4.4.2
decouple==0.0.7
defusedxml==0.6.0
dill==0.3.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-utils-six==2.0
djangorestframework==3.12.4
dnspython==1.16.0
easy-thumbnails==2.4.1
entrypoints==0.3
fake-useragent==0.1.11
filelock==3.0.12
Flask==2.0.1
flatbuffers==1.12
fsspec==2021.4.0
future==0.16.0
gast==0.3.3
geckodriver-autoinstaller==0.1.0
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
huggingface-hub==0.0.8
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonfield==3.1.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
lxml==4.6.1
Markdown==3.3.4
MarkupSafe==2.0.1
mistune==0.8.4
msedge-selenium-tools==3.141.3
multiprocess==0.70.11.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.5
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==20.4
pandas==1.2.3
pandocfilters==1.4.3
parsel==1.6.0
parso==0.7.1
pathspec==0.5.9
pickleshare==0.7.5
Pillow==8.3.1
ping3==2.9.1
pip-upgrade==0.0.6
pipenv==2020.11.15
pipfile==0.0.2
plotly==5.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.15.8
psycopg2-binary==2.9.1
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.7.2
PyHamcrest==2.0.2
pymongo==3.11.3
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.2
reportlab==3.5.68
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
selenium==3.141.0
semantic-version==2.8.5
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.1
sqlparse==0.2.3
tenacity==7.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf==1.0.0
threadpoolctl==2.1.0
tokenizer==3.1.0
toml==0.10.2
tornado==6.1
tqdm==4.49.0
traitlets==5.0.5
TwoCaptcha==0.0.1
typing-extensions==3.7.4.3
ua-parser==0.10.0
Unipath==1.1
urllib3==1.26.2
user-agents==2.2.0
useragent==0.1.1
virtualenv==20.1.0
virtualenv-clone==0.5.4
w3lib==1.22.0
wcwidth==0.1.9
webdriver-manager==3.2.2
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==3.3.0
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
xlwt==1.3.0
xxhash==2.0.2
zipp==3.4.0
zope.interface==5.2.0